requirements.txt
python-dateutil
sympy
